add additionalTarget and update docker-build workflow
Adds an additionalTarget optional param to docker-build workflow. If set, the workflow
will build that docker stage in addition to the default, and tag it with the default tags plus
a dash and the name of the target.

Also updated to the latest versions of the actions used and cleaned up the code. Switched
to docker/metadata-action for generating initial tags and labels.